requirements.txt
asgiref==3.7.2
Django==4.2
django-appconf==1.0.6
django-imagekit==5.0.0
django-mdeditor==0.1.20
django-uuslug==2.0.0
Markdown==3.5.1
pilkit==3.0
Pillow==10.1.0
pip==23.3.2
PyMySQL==1.1.0
python-slugify==8.0.1
setuptools==68.2.2
sqlparse==0.4.4
text-unidecode==1.3
typing_extensions==4.9.0
wheel==0.42.0
